PIM-6442: Move value factories to `Value` namespace Move from Pim\Component\Catalog\Factory\ProductValue to Pim\Component\Catalog\Factory\Value
