+++ src/Pim/Component/Catalog/spec/Factory/Value/PriceCollectionValueFactorySpec.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace spec\Pim\Component\Catalog\Factory\ProductValue;
+namespace spec\Pim\Component\Catalog\Factory\Value;
 
 use Akeneo\Component\StorageUtils\Exception\InvalidPropertyTypeException;
 use PhpSpec\ObjectBehavior;
 use Pim\Component\Catalog\Factory\PriceFactory;
-use Pim\Component\Catalog\Factory\ProductValue\PriceCollectionValueFactory;
+use Pim\Component\Catalog\Factory\Value\PriceCollectionValueFactory;
 use Pim\Component\Catalog\Model\AttributeInterface;
 use Pim\Component\Catalog\Model\PriceCollection;
 use Pim\Component\Catalog\Model\ProductPriceInterface;
